add working app

+++ app/src/main/java/com/example/weatherator/data/models/weather_new/Maximum.kt
@@ -3,5 +3,5 @@ package com.example.weatherator.data.models.weather_new
 data class Maximum(
     val Unit: String,
     val UnitType: Int,
-    val Value: Int
+    val Value: Double
 )
